Move userspace context to own module.
